Consolidated TrueAnal/Swallowed/Nympho into a single file: networkSteppedUp.py
