+++ Contents/Code/networkFPN.py
@@ -16,6 +16,8 @@ def posterAlreadyExists(posterUrl,metadata):
 def search(results,encodedTitle,title,searchTitle,siteNum,lang,searchByDateActor,searchDate, searchAll, searchSiteID):
     searchResults = HTML.ElementFromURL(PAsearchSites.getSearchSearchURL(searchSiteID) + encodedTitle)
     for searchResult in searchResults.xpath('//div[@class="update animation-element bounce-up"]'):
+        if searchSiteID != 9999:
+            siteNum = searchSiteID
         titleNoFormatting = searchResult.xpath('.//a[@class="title"]')[0].text_content().strip()
         Log("Result Title: " + titleNoFormatting)
         curID = PAsearchSites.getSearchSearchURL(searchSiteID) + titleNoFormatting.lower().replace(' ','+')
